Added more variety to jungle trees and conifers shapes
Also improved the look of both a little bit, and re-balanced the spawn/growth
chances, and reduced the wait time sicne we're not using a ton of code to do
this anymore.

Also made the spawn routine clear out any leaves that are in the way of a
spawning tree's trunk, so the trunks don't look like they've been cut by
leaves.