Multiple changes, see below:
Fix replacement of default trees so that it works correctly.  Beeches
now use default tree nodes, but L-systems models.  This redefines the
default tree, wood, leaves, and saplings so that they use moretrees'
textures "Beech" in their descriptions, but only if replacement of
default trees is enabled.

Added backward-compat aliases for this change.

Moved tree replacement mapgen->air aliases to a more logical spot in the
code.

Also, changed out all leaf decay dig_node() calls for nodeupdate(), since the
former calls the latter anyway, and the latter is what does the job.

Increased the palm leaf decay radius to 12.