Remove failed attempt at manual placement flag
The code was failing to set the "placed" flag on empty rubber tree trunk
nodes because the after_place_node callback that does it would only be
called by the default on_place callback, which is overridden for that
node type.  The code also fails to even attempt to set the flag on full
rubber tree trunk nodes.  The technic mod no longer looks at this flag,
so remove the attempt to set it, rather than fix setting it.