Restore previous pine as cedar
Commit ad0cbbc0f disabled the previous moretrees pine, and aliased it to the
default minetest pine.  The default pine is a completely different tree though,
with different apppearance and different textures.

This commit restores the previous pine tree, renaming it to cedar to avoid a tree
name clash with the default game. Any existing moretrees pine trees will be
converted to cedar, so that their appearance does not change. Moretrees pine
trees that have already been converted to the default pine tree are not converted
back.