Add date palm
Spawns in hot and dry desert, but requires water. This makes it a bit harder to
find, but where found, it can be more abundant.
If found in the middle of a desert, its presence indicates a water source below
the surface.

As an additional optional feature, dates (slowly) grow back after harvesting.