Add recipes with dates (and nuts, and coconut)
Using a combination of dates, nuts and coconut, an extremely high
energy date-nut cake can be baked (32 units of food). It's main purpose
is to be a very concentrated transport mechanism of food, but it can be
used for instant replenishment of a starving soul. In practise, it will
be used to craft date-nut energy bars, which merely have a good
nutritional value (4 units).

Because of its exceptional nature, the cake requires special ingredients
(all three types of nuts, as well as coconuts), and given the ingredients,
3 steps are needed (including one cooking step) to bake it. All of the
ingredients themselves, except dates, are not found in nature, but must
also be prepared.