Add new screenshot featuring all trees & instructions for recreating it
The previous screenshot showed only a few of the available trees.
The new one shows all of them (except acacia, which is disabled,
and in the default game anyway, and beech).

As finding a suitable spot where even a significant proportion of
trees is only visible at all is near impossible, this scene was
crafted, and instructions and a script for recreating it are included.